backend/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.23.2

# Video processing
opencv-python-headless==4.8.1.78
av==11.0.0

# WebRTC and async
aiortc<11.0.0
aiohttp==3.8.6
websockets==12.0

# Other utilities
pytz
numpy==1.26.1
requests
pyyaml